third_party/blink/web_tests/external/wpt/app-history/ordering/back-same-document.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<meta name="variant" content="?">
<meta name="variant" content="?transitionWhile">
<meta name="variant" content="?currentchange">
<meta name="variant" content="?transitionWhile&currentChange">

<script>
const variants = new Set((new URLSearchParams(location.search)).keys());

promise_test(async t => {
  // Wait for after the load event so that the navigation doesn't get converted
  // into a replace navigation.
  await new Promise(resolve => window.onload = () => t.step_timeout(resolve, 0));

  await appHistory.navigate("#1").finished;

  const events = [];
  appHistory.addEventListener("navigatesuccess", () => events.push(`navigatesuccess ${location.hash}`));
  appHistory.addEventListener("navigateerror", () => events.push(`navigateerror ${location.hash}`));
  if (variants.has("currentchange")) {
    appHistory.addEventListener("currentchange", () => events.push(`currentchange ${location.hash}`));
  }

  if (variants.has("transitionWhile")) {
    appHistory.addEventListener("navigate", e => {
      e.transitionWhile(Promise.resolve());
    });
  }

  const { committed, finished } = appHistory.back();
  committed.then(
    () => events.push(`committed fulfilled ${location.hash}`),
    () => events.push(`committed rejected ${location.hash}`)
  );
  finished.then(
    () => events.push(`finished fulfilled ${location.hash}`),
    () => events.push(`finished rejected ${location.hash}`)
  );

  Promise.resolve().then(() => events.push(`promise microtask ${location.hash}`));

  await finished;

  if (variants.has("currentchange")) {
    assert_array_equals(events, [
      "promise microtask #1",
      "currentchange ",
      "committed fulfilled ",
      "navigatesuccess ",
      "finished fulfilled "
    ]);
  } else {
    assert_array_equals(events, [
      "promise microtask #1",
      "committed fulfilled ",
      "navigatesuccess ",
      "finished fulfilled "
    ]);
  }
}, "event and promise ordering for same-document appHistory.back()");
</script>
